Editor: split out base class for Script Editing pane
The ScriptEditorBase parent class implements menu commands shared between standard ScriptEditor and DialogEditor.
Copied code mostly from the ScriptEditor, as its list of commands is more full.
Few commands that are not applicable in DialogEditor (at least not yet) are added in override methods of ScriptEditor.
Moved common operations (search, goto line, etc) to ScriptEditorBase.
