+++ Editor/AGS.Editor/AGSEditor.csproj
@@ -221,6 +221,9 @@
     <Compile Include="Panes\Room\RoomEditFilters\CharactersEditorFilter.cs" />
     <Compile Include="Panes\Room\RoomEditFilters\DesignTimeProperties.cs" />
     <Compile Include="Panes\Room\RoomEditFilters\RoomFilterContextMenuArgs.cs" />
+    <Compile Include="Panes\ScriptEditorBase.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
     <Compile Include="Panes\ZoomTrackbar.cs">
       <SubType>UserControl</SubType>
     </Compile>
